move image extraction integration test into extract.test.ts
